refactor(warp): improve type flexibility and add utility functions
Replace specific Warp type annotations with generic `Any` to enhance code flexibility and reduce dependency on internal Warp types. Add `from_jax` utility function for converting JAX arrays to Warp arrays with support for vector and matrix types. Improve documentation with detailed docstrings across elastic energy utility functions. Enhance testing by adding support for both float32 and float64 data types and introducing new tests for Hessian quadratic forms.

These changes improve interoperability between JAX and Warp, increase code maintainability through better documentation, and ensure robust testing across different precision modes.

+++ src/liblaf/apple/warp/sim/energy/elastic/arap/func.py
@@ -9,7 +9,7 @@ from liblaf.apple.warp.typing import float_, mat33, mat43
 
 @wp.struct
 class Params:
-    mu: float_
+    mu: float
 
 
 @wp.func
